requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
pydantic>=2.9.0
streamlit>=1.31.0
pandas>=2.0.0
